Add the "shared" project to hold shared test code and restructured some tests with it
