+++ input/src/main/scala/fix/ExpandRelativeQuotedIdent.scala
@@ -5,11 +5,9 @@ OrganizeImports.expandRelative = true
 
 package fix
 
-import ExpandRelativeQuotedIdent.`a.b`
+import QuotedIdent.`a.b`
 import `a.b`.c
 
 object ExpandRelativeQuotedIdent {
-  object `a.b` {
-    object c
-  }
+  val refC = c
 }
