+++ output/src/main/scala/fix/GroupedImportsMergeUnimports.scala
@@ -1,0 +1,5 @@
+package fix
+
+import fix.MergeImports.Unimport.{b => B, c => _, _}
+
+object GroupedImportsMergeUnimports
