+++ shared/src/main/scala/fix/MergeImports.scala
@@ -1,0 +1,35 @@
+package fix
+
+object MergeImports {
+  object Wildcard1 {
+    object a
+    object b
+    object c
+    object d
+  }
+
+  object Wildcard2 {
+    object a
+    object b
+  }
+
+  object Unimport {
+    object a
+    object b
+    object c
+    object d
+  }
+
+  object Rename1 {
+    object a
+    object b
+    object c
+    object d
+  }
+
+  object Rename2 {
+    object a
+    object b
+    object c
+  }
+}
