+++ shared/src/main/scala/fix/QuotedIdent.scala
@@ -1,0 +1,7 @@
+package fix
+
+object QuotedIdent {
+  object `a.b` {
+    object c
+  }
+}
